добавление записи в отчет по собакам

+++ src/main/java/com/example/telegrampetbot/repositories/ClientRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface ClientRepository extends JpaRepository<Client, Integer> {
-    Optional<Client> findById(Integer id);
+    Optional<Client> findById(Long id);
     Client findByChatId(Long chatId);
 
 }
